+++ docs/updates/CHANGELOG.md
@@ -38,6 +38,10 @@ This file keep's track of all change's made to the operating system during updat
 
 ## Version 0.2
 
+**Security:**
+
+- Panic system (BSOD).
+
 **Tools:**
 
 - Multi make build system.